+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 # CHANGELOG
+## **WORK IN PROGRESS**
+- [ ] delete per day
+- [ ] save per day
+- [ ] archive entries
+
 ## 0.6.2 (2022-11-29)
 - fix day shifting
 - enable setting last_in_block